backend/requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.21.0
nltk==3.6.6
packaging==20.3
pandas==1.2.3
pep517==0.8.2
Pillow==9.0.0
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
regex==2021.3.17
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.4.2
torch==1.8.0
torchvision==0.9.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.5
webencodings==0.5.1
Werkzeug==1.0.1
